requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements-dev.txt

-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
hypothesis>=6.0.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Type stubs
pandas-stubs>=2.0.0
types-PyYAML>=6.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
